Truncate overlong name and displayName values
Previously, we just returned an error if a name or displayName value for
a PublicKeyCredentialEntity was longer than the supported 64 bytes.
With this patch, we instead truncate the value at a UTF-8 character
boundary.  The logic for determining the truncation point is borrowed
from a nightly function from the standard library.

Fixes: https://github.com/Nitrokey/ctap-types/issues/4